Rewrote verification logic to assume all nodes share the same key
After a meeting with Ken, we decided that all nodes in the same Derecho
group would share the same private key for signing log updates, and
unique keypairs would only be used for different georeplication sites.
This makes it much easier to verify another replica's update even with
the signature chains, since the remote node and the local node should
both have the exact same "previous signature" in their logs. It also
means there's need to keep track of a public key for each node in the
group or exchange public keys when a node joins.

The procedure for verifying the signature each replica posts to the SST
is now completely implemented and appears to work. Also, I realized that
the names of the methods I added to Persistent were inconsistent with
the existing methods, so I changed the new signature-related methods to
use CamelCase.