Replaced const version_t& with version_t, since it's an integer
Primitive integer types are usually passed by value in function
parameters, not by const reference, because a pointer to an integer is
at least as large as the integer itself (so passing a constant pointer
to an integer doesn't save any work over copying the integer). For some
reason many functions took version_t parameters by const reference even
though version_t is an integer, so I changed them to take version_t
values instead. The only exception is the trim() function, which is
actually a template that could match either version_t or HLC and so
needs to be a const reference for the case where its parameter is HLC.