Added a callback for global "signature stability", fixed bugs
I added a new user-facing callback for global signature stability (when
an update is verified by all replicas) mirroring the ones for global
stability and persistence. In the process I refactored MulticastGroup a
little to make the predicates and callbacks easier to read.

I also noticed and fixed some bugs in how PersistentRegistry handled
signature chains when versions were not consecutive.