Fixed more bugs in signature logic, added another test
Further testing revealed another problem with how the special case of
the "genesis signature" was handled: since the very first entry would
have a "previous signed version" of INVALID_VERSION, attempting to get
the signature on that entry would appear to always fail.

The processEntryAtVersion method in FilePersistLog contained a serious
bug: The data it supplied to the processing function (i.e. to update the
signer or verifier) was the log header's bytes, not the log data, even
though the provided "buffer size" was fields.dlen and the intention was
clearly to process the actual data.