Replaced PersistentObjectFunctions with an equivalent abstract class
Instead of wrapping all the member functions of a Persistent<T> in
std::functions in order to call them without knowing the template
parameter T, we could equivalently give Persistent<T> an abstract base
class that defines all those functions as virtual. The cost of a virtual
function call is very similar to the cost of an indirect function call
through a std::function wrapper, and doing it this way eliminates the
need to redundantly write down all the method names again in the
register_callbacks() method (when binding them to the std::functions).

I also tweaked the method and member names in PersistentRegistry to be
consistent with the naming convention in the rest of the Persistent
classes: they should all be camelCase, not snake_case, and member
variables are all named with the prefix "m_".