Moved persistent_latency_test out of archive
Similar to my efforts with persistent_bw_test, I revised this old test
to use the refactored PartialSendersAllocator, and use a workaround
using the stability callback to correctly stop when the last version is
persisted (now that the last persisted version is not reliably
total_num_messages - 1). I didn't have time to also make it average its
data and write out a file, though.