Made signatures-enabled configuration separate per subgroup
Instead of a single configuration option that enables signatures for the
entire group, I added a constructor argument to Persistent that enables
signatures for just a single Persistent object, and added a marker
interface for Replicated<T> that you inherit to indicate that you
constructed your Persistent fields with signatures enabled. This allows
a group to have some subgroups that use signatures and some subgroups
that don't. I also updated signed_log_test to test a setup with mixed
signed and unsigned subgroups.

Note that, because of how PersistenceManager works, this assumes that
any Replicated<T> that inherits SignedPersistentFields will construct
all of its Persistent fields with signatures enabled. It will probably
break if you inherit SignedPersistentFields but then construct a
Persistent<T> field with enable_signatures = false.