Changed signed_bw_test back to GB/s and improved bytes_object
The Bytes object used in this and other tests had a from_bytes_noalloc
implementation that actually did an unnecessary memcpy into a new byte
array (it created a new Bytes object from the contents of the
serialization buffer). I wrote a version that eliminates this copy by
making a Bytes object that points directly into the serialization
buffer, which is more in the spirit of from_bytes_noalloc.