Fixed CompletionTracker bug in the signed store mockups
I overlooked the fact that the global persistence callback can be called
once for a batch of versions, which means CompletionTracker needs to
notify the whole batch of futures (not just the version that was passed
to the callback).

Also, I gave up on the "batch" function in signed_store_test, and just
called the single update function in a loop in main. This means there's
no way to detect the intermediate completion times, but at least it
works.