Persistent<T>::getDelta(ver) now returns the delta of EXACT version.
The default behavior of Persistent<T>::get(ver) will return the state of T at version 'ver'. If there is no log exists for 'ver', its latest state before version 'ver' will be returned, which represents the state of the object at a virtual virtual 'ver'. However, although Persistent<T>::getDelta(ver) share a lot of logic with Persistent<T>::get(ver), it has a different semantic: it needs to return the delta at exact 'ver', which does not exists in the given example. This patch introduced a boolean 'exact' flag in the parameter list of PersistLog::getEntry() and PersistLog::getVersionIndex(), so that the Persistent<T>::getDelta(ver) will throw an exception on an invalid version by enforcing the flag.