Finished implementing changes for #178
Now that the macros are working correctly, I was able to trace down and
fix my other bugs. Namely, the return type of tag_ordered and tag_p2p
needed to include the to_internal_tag function (not just the actual call
to return), and I needed to implement a non-const tag_p2p even though
it's a bad idea because the current implementation of Persistent<T>
can't be read from a const method.

All tests have also been updated to use the new syntax for registering
RPC functions, so a simple "make" should complete successfully.