Added more const-correctness to P2P functions
This required adding one more const decoration to Persistent<T>, on
getSignature(), and adding a const version of the operator* shortcut to
get the current version. Almost all P2P functions are const now, except
for the ones that need to modify purely-local, non-replicated state like
a random number generator for the load-balancer.

Also, removed derecho_caller_test, since it's no longer a good example
of how to use Replicated RPC functions and we have equivalent tests for
that functionality.