Fixed errors in ViewManager's use of SST put()
As pointed out by Ken in an e-mail, there were several parts of
ViewManager's code that used a series of single-field SST put()s, with
the assumption that the final put() would arrive on the remote nodes
last, without using a put_with_completion() call to wait for earlier
put()s to complete. I replaced the second-to-last put() call with a
put_with_completion() in all cases where the last put() call needed to
be strictly "fenced" from the earlier ones.