Initial implementation of the framework needed for #167
This is a hopefully-complete basic implementation of the changes
necessary in order to allow callers of ordered_send to wait for their
update to finish persisting without resorting to the single global
persistence callback (as described in #167). I added std::futures to
QueryResults representing local persistence and global persistence being
done, and fulfilled them using corresponding promises in PendingResults
at the points where the existing persistence callbacks would be called.
This required some reworking of internal plumbing with callbacks, and in
the process I renamed the CallbackSet struct UserMessageCallbacks, to
clarify that it was for user-facing callbacks and shouldn't be modified
internally.