Moved multicast_throughput out of archive into unit_tests
This looks like a nice stand-alone test of the SST Multicast system that
was probably written back when SMC was first developed. It's written
like a performance test, but I'm putting it in unit_tests because it
tests one small module of the Derecho system rather than the performance
of the system as a whole.