Moved local persistence callback earlier in handle_persist_request
There seems to be a race condition between the local persistence
callback, which is fired by the persistence manager thread, and the
global persistence callback, which is fired by the SST predicates
thread. Calling the local persistence callback before updating
persisted_num in the SST should ensure that the local persistence
callback always happens before the global persistence callback for a
given update (which is the ordering you would expect).