Added a std::future for the signature verification event
Using the same framework as the futures for local and global
persistence, I added a future to PendingResults/QueryResults
representing the "global verification" event (all replicas have valid
signatures) that is triggered by an internal callback to RPCManager.
This allows the signed_store_test and signed_store_mockup programs to
eliminate the "CompletionTracker" crutch entirely, and probably finishes
fixing issue #167 (it should still be tested, though).