Initial mockup of client-side callbacks via P2P messages
The client_callback_mockup program is an attempt to allow nodes outside
a subgroup to request callbacks when updates in that subgroup reach
persistent or signed states, via P2P messages sent over the existing RPC
framework. If this works, the feature could be baked into Derecho so
that any client node could use it without needing to implement the
receive_callback() and register_callback() functions.