Fixed a bug in signature-verification futures. Closes #167.
Now that users can wait for a signature to be verified by using the
future in QueryResults, a client may be interested in signature
verification status even if it doesn't register a global verification
callback. This if statement in MulticastGroup setup is no longer
appropriate.

With this fix, all the changes for #167 have been implemented and
tested.