Fixed a bug in RPCManager's handling of P2P requests
Since p2p_request_worker runs in a while(!thread_shutdown) loop, but the
local variable reply_size is declared outside the loop, reply_size is
only initialized once and then gets reused on every iteration of the
loop. This means if(reply_size > 0) will not correctly detect whether
receive_message populated the reply buffer - if no reply was generated,
reply_size will be left at whatever value it had during the last loop
iteration. This means the "hack" in the else clause does not necessarily
get to run if the function did not generate a reply.