A possible solution for the deadlock in #195
Working with Weijia, I sketched out a new way to manage the P2P
connections in RPCManager using fine-grained locking on a fixed-size
array of P2PConnection pointers. This will probably work, but one
concern is that iterating over all of the connections for probe_all and
check_failures_loop will now be too expensive because of all of the
extra lock operations. We might need to import a concurrent hash map
library after all.