Fixed another bug discovered while testing the fix for #198
The allocator functions (lambdas) in RPCManager were not correctly
testing whether the amount of memory requested exceeded the buffer they
were about to return. In two cases, the reply size was compared to a
value that was *not* equal to the P2P buffer that would be requested
from P2PConnectionManager, which meant the function could return a
buffer pointer even when the reply was too large and would write past
the end of the buffer.

Also, I had previously forgotten to copy my changes from RPCManager over
to ExternalGroup, so I did that.