Added a mechanism to return exceptions over RPC
If an RPC call triggers an exception that respects the std::exception
hierarchy, the node that received the call will serialize and send back
the type name and what() message of the exception so that the caller can
display it when it throws remote_exception_occurred. This requires
enforcing a small "minimum reply size" so that this message can be sent
in an RPC reply no matter what buffer size has been configured.