Another critical bugfix for API change from libfabric 1.7.o to libfabric 1.12.1
In libfabric v1.12.1, verbs provider's fi_cq_open() API does not pick a valid size if given size is zero. In such a case, fi_msg() will always return -FI_EAGAIN, causing an infinite loop in RDMC initialization. The TCP provider is not affected. Instead of let fi_cq_open() to pick one for us, we set it to a fix number 2097152.