Fixed a bug in mutils-serialization's from_bytes_noalloc
The from_bytes_noalloc functions for ByteRepresentable types were
declared ambiguously. The "const" version had a "v" parameter with type
char const* const, while the "non-const" version had a "v" parameter
with type const char*. These types are actually the same from the
caller's perspective, since both could be matched by an argument of type
const char* (the trailing const applies only to the parameter inside the
method), so attempting to use from_bytes_noalloc would result in a "call
of overloaded function is ambiguous" compile error.

To fix this, I changed the "non-const" version of the function to have a
parameter of type char* (with no consts at all). This matches the
declaration of from_bytes_noalloc for POD types below.