Added tests for JSON-based layout, and fixed a bug
I made some general improvements to subgroup_function_tester and added a
new test case to test how the JSON-based layout with reserved node IDs
works. In the process, I discovered and fixed a bug in the new
JSON-based layout system: It did not account for the identical_subgroups
flag and would crash if the number of ShardAllocationPolicy objects was
not equal to num_subgroups.

I also moved container_template_functions.hpp to the utils folder, and
added spdlog/fmt/ostr.h to our "standard" set of includes (from
logger.hpp), so that logging functions in our code can print out sets
and vectors.