Fixed a bug in RPCManager introduced by 163750226994a084b52060ebe5f9393d4e6fc63c
I can't believe I didn't notice this earlier, but the iterator increment
in new_view_callback needs to be outside the if statement now that there
is no conditional branch that could erase from the map (instead of
iterating). This would cause an infinite loop if these maps were ever
non-empty during a view change.