Initial implementation of a solution for #205
As discussed in the GitHub issue, I replaced the array of PendingResults
objects in RemoteInvoker with heap-allocated PendingResults objects that
are tracked via weak_ptrs. Invocation IDs are now the literal memory
address of a weak_ptr to a PendingResults object, which receive_response
can use to find the PendingResults it needs to update. RPCManager uses
a different weak_ptr to deliver reply maps and persistence events. The
only "strong" pointer to a PendingResults object is a shared_ptr inside
its paired QueryResults, which means the PendingResults will get deleted
as soon as the user thread discards the QueryResults.