Fixed RemoteInvoker to not leak memory for void functions
Functions with a return type of void will never send responses, so they
don't actually need a weak_ptr<PendingResults> and in fact will leak
this memory if they allocate it. The RPC request message still needs an
"invocation ID," but it can be a null pointer because it will never be
used to match up replies to requests.

If we ever decide to start sending remote_exception_occurred messages
for void RPC functions, though, this will need to be fixed, because now
there is no way to receive any kind of response to a void function (even
an exception).