Changed RemoteInvoker's weak_ptr<PendingResults> to a shared_ptr
Following the discussion on #205, this should eliminate a memory leak in
the common case where the client's QueryResults goes out of scope before
all the RPC replies have arrived; now receive_response will always be
able to check PendingResults's all_responded and delete its
heap-allocated pointer. On the other hand, this introduces an
uncommon-case memory leak: if a node fails while RemoteInvoker is still
waiting on a reply from it, receive_response will never be able to
delete the shared_ptr corresponding to that reply, and the
PendingResults will be leaked.