A possible fix for the leaked PendingResults on node failure
After a meeting with Weijia, I realized that storing a copy of the
heap-allocated shared_ptr<PendingResults> in the PendingResults itself
would allow RPCManager to delete the shared_ptr if RemoteInvoker loses
track of it. When RPCManager delivers a set_exception_for_removed_node
event to a PendingResults object, it can be certain that RemoteInvoker
won't be getting any more replies for that object, and the shared_ptr
should be deleted. Since RemoteInvoker can't delete the shared_ptr if it
doesn't get any more replies for it, RPCManager deletes the shared_ptr
through the PendingResults (essentially, it asks PendingResults to
delete itself). See #205.