Fixed another bug with PendingResults cleanup
It turns out that it's possible to receive an RPC reply from a
non-failed node for a message that was delivered in a previous view. The
view change logic only ensures that the (ordered) RPC *request* messages
are delivered before moving to the next view, and in fact the
reply-listening thread (p2p_receive_loop) remains blocked during a view
change, so any replies triggered by an RPC message that was delivered
during ragged edge cleanup will not be handled until after the view
change is over. This means that RPCManager's new view callback should
*not* delete a PendingResults just because some of its destination nodes
were removed; the non-failed nodes could still deliver replies to that
PendingResults after the view change finishes.

Also, added some configuration options to the default logger to make it
easier to find these bugs.