Fixed a silly use-after-free error in PendingResults
Testing self_heap_ptr for null, and setting it to null after deleting
it, won't work if self_heap_ptr is actually the last valid shared_ptr to
the PendingResults - then delete self_heap_ptr will delete the very
PendingResults that stores self_heap_ptr, and it will be illegal to set
self_heap_ptr to null! The call to delete self_heap_ptr must be the last
statement in the method, since the object it's in might get deleted at
that point. We can still prevent an accidental double-free by just using
a boolean that gets changed before deleting the pointer.