A partial fix for #213
Some nodes were moving to the next view as soon as they detected the first
change in a batch of 3 changes, and thus only updated their num_acked and
num_committed by 1, even though the view they then installed included all 3
changes in the batch. This caused them to immediately detect a new
"unacknowledged" change in the next view, which was actually one they
already installed. The problem was with the leader_committed_changes
predicate, which should be checking to make sure sure the leader has finished
proposing all the changes in a batch before returning true -- the call
to changes_includes_end_of_view was supposed to do this, but it only
checks to see if the "end-of-view" change is in the changes list, not
whether this node has actually acknowledged it yet. I fixed this by
making leader_committed_changes call a new function called
end_of_view_committed instead.

Unfortunately some nodes can still get stuck after a rapid succession of
view changes, but it's no longer because they spuriously detect a new
change proposal.

Note: I also made some cosmetic changes to aid debugging, like adding more
comments and rearranging the code in register_predicates to make it more
obvious which predicates are paired with which triggers.