Made RPCManager keep track of external client IDs
Some earlier problems were caused by the fact that RPCManager's
report_failure method needed to call view_manger.get_members() in order
to determine whether the failed node was an internal or external node,
and get_members() needs to acquire the view_mutex. However, there is
actually no reason why checking for an external client should need to
acquire the view mutex or check the current view's members, because
external clients do not join or leave with view changes and do not
affect the current view. Instead, RPCManager can just maintain a
separate list of external clients, which does not need to be shared with
ViewManager and does not need to be synchronized. ViewManager's
external_join_handler already makes sure that external clients won't
have the same ID as an internal node.