Minor improvements to make shutdown more graceful
The leader still aborts abnormally when attempting a graceful shutdown,
because it detects that "some other node marked this one as failed" even
though that "other node" is itself. I think the point of the RIP field
is to check whether a node is intentionally shutting down, so I should
check for this in propose_changes.