Renamed ExternalGroup to ExternalGroupClient to reduce confusion
This object doesn't represent a "group," it represents a single client
of a group. Hopefully this will make it easier to reason about the
behavior of external clients and what the template parameters mean.

Also added a bunch of documenting comments and tried to make the
client_callback_mockup files easier to understand.