Added some new methods to get signatures from a Delta entry
The GetDeltaSignature method takes a "search predicate" that it will use
to determine if the log entry at the requested version is the "right
one" without needing to deserialize and return that log entry. This
should allow me to implement a get_signature() method for
SignatureCascadeStore that works similarly to the get() method for the
regular PersistentCascadeStore.

I'm not sure how to test these on their own, though, since nothing in
the Derecho codebase implements a delta with embedded keys, like the
DeltaCascadeStoreCore does. I might have to implement a mockup of
DeltaCascadeStoreCore in the applications/tests/unit_tests folder.