Added some new utility methods to the Group interface
For my work in Cascade, I need a way to find out how many subgroups of
the same type there are and which "subgroup index" the current node is
in. All our previous applications used only one subgroup of a given
type, so subgroup indexes never came up, but it turns out the Group
interface is hard to use if you don't know your subgroup index.

Also added more documentation to the Group interface (since it's public
facing), and improved our tests to use better methods of determining
which subgroup the node is in.