Added a convenience constructor to Persistent<T>
The way most Derecho applications use Persistent<T> requires them to
write several "boilerplate" arguments for the constructor in order to
specify a value for the PersistentRegistry, since you can't use default
arguments for parameters that come before the non-default one. Instead
of trying to rearrange the parameters so that PersistentRegistry comes
earlier (which might break other use cases relying on the current
default-arguments order), I just wrote a new constructor that delegates
to the real constructor.