Fixed include paths so Derecho can be rebuilt locally
If Derecho is already installed on a system path, and you attempt to
rebuild it locally (in the development directory) after making changes,
you can get missing-definition errors because the compiler is using the
installed Derecho headers rather than the local ones that you made
changes to. It turns out that this behavior is caused by writing
includes with the syntax #include <derecho/core/group.hpp>, which
prefers system-installed headers over local ones. Replacing all the
Derecho includes with the other syntax, #include "derecho/core/group.hpp",
makes the compiler prefer local headers over system-installed ones. The
path in quotes does not need to be a relative path (like
"../../derecho/core/group.hpp") as long as the local include directory
is part of CMake's target_include_directories for each target.