Fixed a bug in signatures feature of Persistent
I somehow overlooked this bug for a long time: most of the signature
operations are no-ops if the Persistent object has signatures disabled,
except updateSignature and updateVerifier. These also need to check if
signature_size = 0 and do nothing if so; otherwise the signer can get
updated with bytes even when signatures should be disabled on a field.