Added ReplicatedType factories to ExternalGroupClient
The register_notification method needs to construct an instance of type
T in order to call build_remote_invocable_class, but not all types T
will have a no-argument constructor. I can work around this by requiring
the user to provide factories for each ReplicatedType, just like in the
Group constructor, although this is unsatisfying because it still means
the ExternalGroupClient is constructing an entire Replicated Object
(which may have a lot of data) just to be able to receive notifications.

While doing this, I also improved make_kind_map to be fully generic. I
guess I finally learned enough black magic to make the templates work.