Created a proper type for Notification messages
As noted in the comments in notification.hpp, client-side notifications
should include an "opcode" header in addition to the byte array. I made
this a separate type, NotificationMessage, rather than reuse the Bytes
object, so that other code can continue to use the Bytes object as it
is. Also, if we're going to make Bytes a part of core Derecho (instead
of just in the tests folder), it should use the better version I wrote
for performance_tests rather than the old version in archive.