Fixed bugs in persistence_notification_test
I forgot that the Bytes object doesn't construct an empty array if you
pass it nullptr. (It did at one point, I think). Added a clearer "empty"
constructor for Bytes, since this is probably a common use case. Also,
bytes_size() can't be used to get the size of the array in Bytes,
because it's for serialization and includes the size of the other fields
(notably size itself), so Bytes needs a size() method like std::vector.