Fixed a bug in persistence_notification_test
The way this code is designed, only the node that received the P2P
update request can detect persistence events for that update, so it
won't work for the client to request a notification from a different
node than the one it sent the update to. That's why it kept crashing.