Experimental support for out-of-order P2P sends
This is a possible way to fix the concurrency bug identified in #217
that does not add any additional locks. Instead of ensuring that a
thread maintains a lock on the P2PConnection between the call to
get_sendbuffer_ptr() and send(), send() requires a parameter identifying
the specific sequence number of the message to send so that it can be
matched up with the buffer received in get_sendbuffer_ptr(). The
get_sendbuffer_ptr() method now increments the outgoing message number
as soon as it returns a buffer, ensuring that no two threads can get the
same buffer pointer, and returns the sequence number to the caller so
that the caller can pass the sequence number back in to send().

Since get_sendbuffer_ptr() and send() are now decoupled, it is possible
to call send() on a higher (later) sequence number before calling send()
on a lower (earlier) sequence number. This ought to work fine because
the receiving side will still receive messages in sequential order (it
will ignore the later message until it gets the earlier message), but I
can't be sure until I test it.