Updated comments to confirm the absence of a bug; close #217
While reviewing the code for my previous change to P2P connections, it
looked like I found an instance where a call to get_sendbuffer_ptr() was
not matched with a call to send(), specifically in rpc_message_handler.
After a meeting with Weijia we determined that this was not a bug after
all, because the RPC reply that a node sends to itself for an
ordered_send does not need its sequence number tracked and does not need
to use the "buffer window" in P2PConnection. I thought I should write
this down so I don't try to debug the same code again in the future.

Weijia's testing has confirmed that issue #217 is now fixed, and no
further changes were needed to my "experimental" commit.