Consistently applied context_ptr<const T> to all from_bytes_noalloc
In an attempt to fix issue #204, I changed the API of all
from_bytes_noalloc functions to match the ones declared for
ByteRepresentable and POD types. Specifically, there are two versions of
from_bytes_noalloc for each type: One that takes a buffer parameter that
is uint8_t const * const and returns a context_ptr<const T>, and one
that takes a buffer parameter that is uint8_t* and returns a
context_ptr<T>. In order to get this to work I had to add const versions
to all the "is_blah" type traits so they would match both T and const T,
and in cases where I couldn't modify the type trait I added
std::remove_cv_t to the template parameter T.

The new unit test deserialize_and_run_test demonstrates that this
resolves the issue, at least at compile-time. The test would not
compile using the previous version of SerializationSupport.hpp, but now
it compiles successfully.