Renamed variables to improve readability
Replaced the inscrutable letter "v" with the word "buffer" throughout
SerializationSupport.hpp whenever "v" was used to mean a pointer to a
buffer. Similarly replaced "f" with "consumer" (since f represents a
buffer-consumer function). Hopefully this will make it a little easier
to use the serialization library, especially for users who want to write
their own serialization functions (instead of using
DEFAULT_SERIALIZATION_SUPPORT) and need to understand the meaning of the
arguments to to_bytes and from_bytes.