Fixed bugs in the ostream implementations
I made that last commit too quickly, without testing it, because I had
to move between computers. Now that I've tested it I see that my loops
didn't handle empty maps/sets correctly because the last item still
needs to be printed explicitly, and the unordered versions need to use a
different trick to check for the iterator reaching the last element.