Fixed bugs in mutils-serialization's tuple and map support
This is mostly a copy and paste of commit
47702f008dde4193606ab30129d29d2398060cd0 in the mutils-serialization
repository. I also copied in the serialization-test.cpp file, so this
repository has a simple way to check for compile errors in
mutils-serialization.

I also added another test to rpc_function_types.cpp so it would test to
make sure you can call an RPC function with a std::map, since this is
the use case that caused me to notice the bugs in mutils-serialization.