Removed obsolete format targets, updated clang-format
clang-format has a new option called SpacesInLineCommentPrefix that
allows you to control whether it forcibly adds spaces after the // in
comments. I realized we should use the option to disable that feature,
since we have a lot of comments with different numbers of leading spaces
and we don't want clang-format to cause a bunch of unnecessary changes.
Also, I realized that the "format_x" targets in our CMakeLists.txt files
don't work any more, because they assume the cpp and header files are
all in the same directory (and also that builds happen in the source
directory).