Added a new test to reveal the bug in persistence with delta
If a Replicated Object has a Persistent field with IDeltaSupport
enabled, and it gets a multicast message that does not update the
persistent-delta field, the persistent-delta field will not create a new
version in its log. This means the persistence_manager thread will hang
on attempting to call Persist on the Replicated Object, because it will
get stuck in the loop in Replicated<T>::persist that loops until
persistent_registry->persist returns the latest version. When the
persistent registry calls persist on the IDeltaSupport field, it will
not return the latest version because it has no log entry for that
version, so the minimum returned by persistent_registry->persist will
continually be less than the latest version.

Also added SO_REUSEADDR to the TCP socket constructor in tcp.cpp because
we expect a restarting Derecho process to reuse the same ports. In
practice if you attempt to restart a Derecho process too quickly it will
fail with "address already in use" because the OS is keeping the port
reserved for the previous (terminated) instance of the Derecho process.