Possible fix for the persistence thread problem
The persist() method will return the version stored in the metadata file
when it is successful, even if that version doesn't exist in the log.
The advanceVersion() method even admits that it can be used to create
gaps in the log, which is what DeltaSupport objects want to do (a gap in
the log when there is no data to record). Maybe calling advanceVersion
is all we need to do.