Fixed an unrelated bug in TCP connections that made it harder to test
The tcp_connections constructor quietly ignored failures to initialize a
connection, which meant that those failures would later be discovered
when an unrelated operation tried to read or write and got an assertion
failure from assert(it != sockets.end()). Instead of trying to batch
initialize connections in the constructor, which is only done in one
place anyways, we should always call add_node one at a time and report
any failures.

Incidentally, the reason TCP connections can sometimes fail at startup
seems to be that when testing locally, Linux may assign a process the
same ephemeral port for its outbound connection as another process is
trying to use as the SST port for its inbound connection, because our
SST/RDMC/GMS ports are in the ephemeral port range. This causes that
process to accidentally connect to itself instead of to the other node
it's trying to connect to, which causes "WARNING: node replied with
wrong id" in tcp_connections.