Fixed the spdlog registry bug
It turns out the reason the RPCLogger I created wasn't in the spdlog
registry after creating it had nothing to do with my code, but with
CMake. Since CMakeLists.txt for the SST module didn't declare a
dependency on spdlog, the spdlog library wasn't getting linked to all
the parts of Derecho. I don't know why this affected RPCManager, which
does correctly declare its dependency on spdlog, but adding the
target_link_libraries directives fixed the problem.