Added module logger for Persistent, finished up RPC logger
There's no single point where we know the first Persistent object will
be created (unlike RPCManager that is definitely initialized before all
RPC objects), so Persistence needs a static factory for the logger
similar to the one for the default logger. Hopefully this is equivalent
to what we were already doing with the dbg_default_debug macros, calling
the static default logger.

Also I noticed a few areas where the log messages needed to go to
rpc_logger but weren't.