Bugfixes for OOB completion entries.
Addressed issue: waiting for OOB completion sometimes caused segmentation fault. How does this happen?
First of all, the OOB completion entries were managed by `std::unique_ptr` and the memory was allocated by `malloc()`. Those pointers were put in a thread local queue which holded the memory before completion events came. Here I assumed the completion events would come in the order of the corresponding operations in the same thread. Therefore, we simply popped the front completion entry of the thread-local queue, without checking if the compeltion entry match what we are waiting for (Another design issue to be addressed later). Although that assumption might be true for send operations, it does not stand when we mix the send (`ibv_post_send`) and receive (`ibv_post_recv`). If the events were out of order, a later completion event held a broken pointer pointing to the memory being released by poping a previous compeltion entry. Here came SIGSEGV!

In this fix, we first change `struct lf_sender_ctxt` to `class lf_completion_entry_ctxt` because both send and receive operations could generate a completion event. We also introduced to it an attribute flag called `_managed` to identify if the memory of `lf_completion_entry_ctxt` object is managed by derecho (and therefore, the completion polling thread does not need to release it) or not (and therefore, the completion polling thread is responsible of release it after processing.) Then, we remove the thread-local queue for the completion entries and use the `managed` flag to notify the completion polling thread of how to handle its memory.

Smoke-tested in this commit. More careful tests are needed.